Moved Rewarding Status to rewards
